split svg-general-components.ts 2/2
